Added backend to app
Using JSON-server allowed me to have a mock backend for this app. This allowed data and any changes to the data to be saved and stored on a backend so nothing is lost if user closes/refreshes page after a change. User may add, delete or change reminder status and any changes will be saved to the back end